+++ MathUtils.h
@@ -12,7 +12,7 @@ public:
 	/// <param name="lowerBound"></param>
 	/// <param name="upperBound"></param>
 	/// <returns></returns>
-	static int Clamp(int value, const int& lowerBound, const int& upperBound);
+	static float Clamp(float value, const float& lowerBound, const float& upperBound);
 
-	static Vector2 ClampVector2Magnitude(Vector2 vector, const int& lowerBound, const int& upperBound);
+	static Vector2 ClampVector2Magnitude(Vector2 vector, const float& lowerBound, const float& upperBound);
 };
